requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
tiktoken==0.5.1
backoff==2.2.1